# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEVIUM AI Writing Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a book in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /projects/{projectId}/books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Chapters ───────────────────────────────────────────────────────────────
  /books/{bookId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters in a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter in a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "201":
          description: Created chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /books/{bookId}/chapters/{id}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChapterBody"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Scenes ─────────────────────────────────────────────────────────────────
...
  /projects/{projectId}/characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List characters in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /projects/{projectId}/characters/{id}:
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterBody"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Locations ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        position:
          type: integer
        status:
          type: string
          enum: [draft, in_progress, completed, archived]
        synopsis:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - position
        - status
        - createdAt
        - updatedAt
    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        position:
          type: integer
        status:
          type: string
          enum: [draft, in_progress, completed, archived]
        synopsis:
          type: ["string", "null"]
      required:
        - title
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        position:
          type: integer
        status:
          type: string
          enum: [draft, in_progress, completed, archived]
        synopsis:
          type: ["string", "null"]

    # ── Chapter ───────────────────────────────────────────────────────────────
    Chapter:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        title:
          type: string
        position:
          type: integer
        summary:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, in_review, ready, blocked]
        wordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - title
        - position
        - status
        - wordCount
        - createdAt
        - updatedAt
    CreateChapterBody:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
        summary:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, in_review, ready, blocked]
      required:
        - title
    UpdateChapterBody:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
        summary:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, in_review, ready, blocked]

    # ── Scene ─────────────────────────────────────────────────────────────────
...
    Character:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [protagonist, antagonist, secondary, minor]
        physicalDescription:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        motivations:
          type: ["string", "null"]
        currentState:
          type: ["string", "null"]
        knowledgeState:
          type: ["string", "null"]
        injuries:
          type: ["string", "null"]
        secrets:
          type: ["string", "null"]
        relationships:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - role
        - createdAt
        - updatedAt
    CreateCharacterBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [protagonist, antagonist, secondary, minor]
        physicalDescription:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        motivations:
          type: ["string", "null"]
        currentState:
          type: ["string", "null"]
        knowledgeState:
          type: ["string", "null"]
        injuries:
          type: ["string", "null"]
        secrets:
          type: ["string", "null"]
        relationships:
          type: ["string", "null"]
      required:
        - name
        - role
    UpdateCharacterBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [protagonist, antagonist, secondary, minor]
        physicalDescription:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        motivations:
          type: ["string", "null"]
        currentState:
          type: ["string", "null"]
        knowledgeState:
          type: ["string", "null"]
        injuries:
          type: ["string", "null"]
        secrets:
          type: ["string", "null"]
        relationships:
          type: ["string", "null"]

    # ── Location ──────────────────────────────────────────────────────────────
...